+++ src/bin/osm-lump-ways-down/cli_args.rs
@@ -69,7 +69,6 @@ pub struct Args {
         long = "tag-filter-func",
         value_name = "FILTER_FUNC",
         conflicts_with = "tag_filter",
-        value_parser=opt_read_file::<tagfilter::TagFilterFunc>,
     )]
     pub tag_filter_func: Option<tagfilter::TagFilterFunc>,
 
@@ -225,26 +224,4 @@ pub struct Args {
     /// Upstream from each end goes only up this many nodes.
     #[arg(long)]
     pub ends_upstreams_max_nodes: Option<i64>,
-}
-
-/// CLI arg parser. If the value starts with @, the rest is assumed to be a filename, the contents
-/// of which are parsed to type T
-fn opt_read_file<T>(arg_val: &str) -> Result<T, String>
-where
-    T: std::str::FromStr<Err = String> + std::fmt::Debug,
-{
-    if let Some(filename) = arg_val.strip_prefix('@') {
-        let val = std::fs::read_to_string(filename)
-            .map_err(|e| format!("Couldn't read filename {}: {}", filename, e))?;
-        T::from_str(&val)
-    } else {
-        let res = T::from_str(arg_val);
-
-        if res.is_err() && std::path::Path::new(arg_val).is_file() {
-            let original_error = res.unwrap_err();
-            Err(format!("Unable to parse {:?}. However that is a filename. Did you mean @{} ? Original Error: {}", arg_val, arg_val, original_error))
-        } else {
-            res
-        }
-    }
 }